+++ src/main/kotlin/net/pwall/json/JSONDeserializer.kt
@@ -35,7 +35,12 @@ import kotlin.reflect.KParameter
 import kotlin.reflect.KProperty
 import kotlin.reflect.KType
 import kotlin.reflect.KTypeProjection
-import kotlin.reflect.full.*
+import kotlin.reflect.full.companionObject
+import kotlin.reflect.full.companionObjectInstance
+import kotlin.reflect.full.functions
+import kotlin.reflect.full.isSubclassOf
+import kotlin.reflect.full.isSuperclassOf
+import kotlin.reflect.full.staticFunctions
 import kotlin.reflect.jvm.isAccessible
 
 import java.time.Duration